get_stream_info/src/requirements.txt
Brotli==1.0.9
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mutagen==1.46.0
psycopg2-binary==2.9.6
pycryptodomex==3.17
pytube==12.1.3
six==1.16.0
websockets==10.4
Werkzeug==2.2.3
yt-dlp==2023.3.4
